+++ src/sql_to_fnc.interfaces.ts
@@ -2,4 +2,5 @@ export interface FieldDefinition {
   field: string;
   type: string;
   notNull: boolean;
+  description: string;
 }